# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Store Intelligence API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface StoreInfo {
  store_id: string;
  name: string;
  city: string;
}

export interface StoreMetrics {
  store_id: string;
  unique_visitors: number;
  conversion_rate: number;
  avg_dwell_ms: number;
  total_revenue_inr: number;
  footfall_today: number;
  staff_count: number;
}

export interface ConversionFunnel {
  store_id: string;
  entered: number;
  browsed: number;
  reached_billing: number;
  converted: number;
  avg_basket_inr: number;
}

export interface HeatmapZone {
  zone_id: string;
  label: string;
  visitor_count: number;
  avg_dwell_ms: number;
  score: number;
}

export interface Anomaly {
  anomaly_id: string;
  type: string;
  severity: string;
  message: string;
  detected_at: string;
}

export interface TrafficPoint {
  hour: number;
  day: string;
  visitors: number;
}