# lib/api-zod/src/generated/types/anomaly.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Store Intelligence API
 * OpenAPI spec version: 0.1.0
 */

export interface Anomaly {
  anomaly_id: string;
  type: string;
  severity: string;
  message: string;
  detected_at: string;
}