# lib/api-zod/src/generated/types/conversionFunnel.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Store Intelligence API
 * OpenAPI spec version: 0.1.0
 */

export interface ConversionFunnel {
  store_id: string;
  entered: number;
  browsed: number;
  reached_billing: number;
  converted: number;
  avg_basket_inr: number;
}