# lib/api-zod/src/generated/types/heatmapZone.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Store Intelligence API
 * OpenAPI spec version: 0.1.0
 */

export interface HeatmapZone {
  zone_id: string;
  label: string;
  visitor_count: number;
  avg_dwell_ms: number;
  score: number;
}