# lib/api-zod/src/generated/types/storeInfo.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Store Intelligence API
 * OpenAPI spec version: 0.1.0
 */

export interface StoreInfo {
  store_id: string;
  name: string;
  city: string;
}