# lib/api-zod/src/generated/types/storeMetrics.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Store Intelligence API
 * OpenAPI spec version: 0.1.0
 */

export interface StoreMetrics {
  store_id: string;
  unique_visitors: number;
  conversion_rate: number;
  avg_dwell_ms: number;
  total_revenue_inr: number;
  footfall_today: number;
  staff_count: number;
}